requirements.txt
fastapi
uvicorn[standard]
pytest
httpx
python-dotenv